сделаль по файлькам (сори, онли объектные, но так уже лучше, поверь мне брат)
