+++ common.h
@@ -1,0 +1,11 @@
+//  Кароч сделал этот файл чтоб в мэйне ты мог подключить чисто его, без миллиона инклюдов и всё работало
+
+#ifndef _COMMON_H_
+#define _COMMON_H_
+
+#include "Linked_list_dot.h"
+#include "Linked_list_line.h"
+#include "Linked_list_rectangle.h"
+#include "Parall.h"
+
+#endif
